cancerscope/requirements.txt
coverage
numpy
scikit-learn>=0.18
keras
scipy
requests
pandas
pyyaml
matplotlib>=2.2.2
pysocks==1.6.8
